+++ src/posting/widgets/request/query_editor.py
@@ -1,3 +1,4 @@
+from rich.text import Text
 from textual.app import ComposeResult
 from textual.binding import Binding
 from textual.containers import Vertical
@@ -37,7 +38,9 @@ class ParamsTable(PostingDataTable):
             row = self.get_row_at(row_index)
             params.append(
                 QueryParam(
-                    name=row[0], value=row[1], enabled=self.is_row_enabled_at(row_index)
+                    name=row[0].plain if isinstance(row[0], Text) else row[0],
+                    value=row[1].plain if isinstance(row[1], Text) else row[1],
+                    enabled=self.is_row_enabled_at(row_index),
                 )
             )
         return params
@@ -56,7 +59,7 @@ class QueryStringEditor(Vertical):
                 VariableInput(placeholder="Value"),
                 button_label="Add parameter",
             ),
-            empty_message="There are no parameters.",
+            empty_message="No query parameters",
         )
 
     @property